feat: make use of `receiveOn` over manual dispatch
This lets the effect implementation take ownership
of the queue that an effect turns on over force dispatching.

Also, move Effect type out into separate file for better separation of
code :).
