+++ Cycle/Sources/Cycle/Core.swift
@@ -1,19 +1,6 @@
 import Foundation
 import Combine
 
-public struct Effect<A> {
-    public let run: (@escaping (A) -> Void) -> Void
-
-    public init(run: @escaping (@escaping (A) -> Void) -> Void) {
-        self.run = run
-    }
-
-    public func map<B>(_ f: @escaping (A) -> B) -> Effect<B> {
-        return Effect<B> { callback in self.run { a in callback(f(a)) }
-        }
-    }
-}
-
 public typealias Reducer<Value, Action> = (inout Value, Action) -> [Effect<Action>]
 
 public final class Store<Value, Action>: ObservableObject {
